requirements.txt
# requirements.txt

flask
google-cloud-aiplatform
google-cloud-storage
vertexai
requests
python-dotenv
python-docx
fpdf2
gunicorn